+++ pennylane/devices/default_qubit_tf.py
@@ -109,8 +109,9 @@ class DefaultQubitTF(DefaultQubit):
 
 
     Args:
-        wires (int): the number of wires to initialize the device with
-
+        wires (int, Iterable[Number, str]): Number of subsystems represented by the device,
+            or iterable that contains unique labels for the subsystems as numbers (i.e., ``[-1, 0, 2]``)
+            or strings (``['ancilla', 'q1', 'q2']``). Default 1 if not specified.
         shots (int): How many times the circuit should be evaluated (or sampled) to estimate
             the expectation values. Defaults to 1000 if not specified.
             If ``analytic == True``, then the number of shots is ignored